requirements.txt
python-dotenv~=1.0.1
langchain~=0.1.14
requests~=2.31.0